Revert "9.26 port work"

+++ TEAM2 ONE MONTH CODE/src/Drivers/Outputs/Buzzer/Buzzer_Driver.h
@@ -9,7 +9,7 @@
 #ifndef BUZZER_DRIVER_H_
 #define BUZZER_DRIVER_H_
 
-void buzzer(uint32_t period,uint8_t duty_cycle);
+void buzzer(uint16_t period,uint8_t duty_cycle);
 
 
 #endif /* BUZZER_DRIVER_H_ */
